Help with tracing by making explicit modules
