+++ Cargo.toml
@@ -47,7 +47,7 @@ nix = { version = "0.30.0", default-features = false, features = ["user"] }
 pin-project-lite = { version = "0.2.16", default-features = false }
 serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
 serde = { version = "1.0", default-features = false, features = ["derive"] }
-thiserror = { version = "2.0", default-features = false }
+thiserror = { version = "2.0", default-features = false, features = ["std"] }
 tokio = { version = "^1.22.0", default-features = false, features = ["macros", "rt"] }
 smol-cancellation-token = { version = "0.1", default-features = false }
 tower = { version = "0.5.2", default-features = false }
@@ -57,7 +57,7 @@ tracing-error = "^0.2.0"
 tracing-tree = { version = "^0.4.0", default-features = false }
 zbus = { version = "5.1", default-features = false, features = ["async-io"] }
 serde_plain = "1.0.1"
-ssip = { version = "0.4", default-features = false }
-ssip-client-async = { version = "0.18.0", default-features = false, features = ["async-io"] }
+ssip = { version = "0.5", default-features = false }
+ssip-client-async = { version = "0.19.0", default-features = false, features = ["async-io"] }
 
 xdg = { version = "2.5.2", default-features = false }
