+++ odilia/src/handlers/mod.rs
@@ -1,0 +1,7 @@
+mod atspi;
+mod commands;
+mod input;
+
+pub use atspi::*;
+pub use commands::*;
+pub use input::*;
